Made a few changes, still havent been successful with getting a notification or a date but corrected a few errors and made some tweaks to the app

+++ Calorie Tracker/Calorie Tracker/Calorie+Convenience.swift
@@ -14,8 +14,8 @@ extension Calorie {
                      timestamp: Date = Date(),
                      context: NSManagedObjectContext = CoreDataStack.shared.mainContext) {
         self.init(context: context)
-        
-        self.calorie = calorie
+
+        self.calorie = Int16(calorie)
         self.timestamp = timestamp
     }
 }
